cmake: raise fuzzy match to 95.7% (cycle 25) (#16879)
* cmake: DrawDiaryBase y0 as folded ternary int (89.53%)


* cmake: CmakeVillageDraw runtime float consts + per-block fctiwz + cursor fixes (90.04%)


* cmake: CmakeVillageDraw CColor temp + table hoist (90.09%)


* cmake: CmakeVillageDraw cursor-base 200, %8 wobble, row>=5 decision (90.15%)


* cmake: direct u8 cast of named a255 defeats int-CSE in CmakeVillageDraw (90.29%)


* cmake: CmakeVillageDraw named-double 568 + hoisted cursorBase (90.47%)


* cmake: CmakeVillageDraw cursorBase ternary (90.48%)


* cmake: DrawCmakeName const-width nameX, alpha reassign, >>1 cursor anim (90.91%)


* cmake: CalcSingCMake done-temps + noPad bool materialization (91.72%)


* cmake: reconstruct GetCmakePadDown/Repeat inline helpers (91.76%)


* cmake: CalcSingCMake case2 shared done=0 via goto (91.78%)


* cmake: CalcSingCMake case5/6 shared done=0 goto (91.80%)


* cmake: CmakeTribeCtrl/CmakeJobCtrl return int (91.81%)


* cmake: CmakeNameDraw a255/u8 recipe + titleW + crestRightX + decision fix (92.34%)


* cmake: CmakeNameDraw cursor 229/200 bases, font loop form, a255 decl order (92.70%)


* cmake: CmakeTribeDraw a255 recipe + boxW/boxH 416x240 title rect (93.09%)


* cmake: CmakeTribeDraw float cursor coords + %8 wobble (93.63%)


* cmake: CmakeSexDraw panel 256x162.46 runtime coords + /2.0 + a255 (93.94%)


* cmake: /2.0 reciprocal-divide form for centering exprs (93.99%)


* cmake: CmakeNameCtrl early-return-0, boundary 5:4, junction temps (94.04%)


* cmake: CmakeNameCtrl repeat/down decl order (94.04%)


* cmake: IsDuplicateCmakeName drop empty guard + uncached slot (94.08%)


* cmake: DrawCmakeDecision a255 recipe (94.12%)


* cmake: DrawCmakeDecision runtime cursorY 373.0 + /2.0 tx (94.32%)


* cmake: DrawCmakeYesNo a255/CColor-temp//2.0f (94.32%)


* cmake: CmakeResultDraw duplicated full-alpha panel block (94.82%)


* cmake: CmakeTribeCtrl junction temps + tribeCount-1 (94.86%)


* cmake: CmakeJobDraw cursorY float roundtrip (95.09%)


* cmake: CmakeNameCtrl reorder + j=i init (95.18%)


* cmake: CmakeNameCtrl cntlzw empty-test idiom (95.21%)


* cmake: CmakeNameCtrl backspace arm order (95.65%)


* cmake: CmakeVillageCtrl junction temps (95.69%)


* cmake: CmakeJobCtrl junction temps (95.72%)


---------

+++ include/ffcc/p_menu.h
@@ -529,11 +529,11 @@ public:
     void CmakeSexClose();
     void CmakeSexDraw();
     void CmakeTribeOpen();
-    unsigned short CmakeTribeCtrl();
+    int CmakeTribeCtrl();
     void CmakeTribeClose();
     void CmakeTribeDraw();
     void CmakeJobOpen();
-    unsigned short CmakeJobCtrl();
+    int CmakeJobCtrl();
     void CmakeJobClose();
     void CmakeJobDraw();
     void CmakeResultOpen();
